requirements.txt
openai
pyairtable
google-genai
requests
schedule
python-dotenv
